chore: rename utils package
