+++ pages/2_Validate.py
@@ -4,11 +4,11 @@
 
 import streamlit as st
 
-from utils.predictions_loader import load_prediction_file
-from utils.schema_inspector import SchemaInspector
-from utils.session_manager import SessionManager
-from utils.styles import VALIDATE_PAGE_STYLES
-from utils.validation_ui import generate_validation_block
+from mesa_validate.predictions_loader import load_prediction_file
+from mesa_validate.schema_inspector import SchemaInspector
+from mesa_validate.session_manager import SessionManager
+from mesa_validate.styles import VALIDATE_PAGE_STYLES
+from mesa_validate.validation_ui import generate_validation_block
 
 st.set_page_config(page_title="Validate", layout="wide")
 st.logo("aic_logo.png")
